reduced heatmap export.

+++ gain_imputation.py
@@ -692,28 +692,29 @@ def main(args, raw_data, original_data_x, ids, timestamp, date_str, ss_data):
     df_ss.columns = header
     dfs_ss = [g for _, g in df_ss.groupby(["id"])]
 
-    pool = Pool(processes=args.n_job)
-    for i in range(len(dfs_transponder)):
-        result = pool.apply_async(
-            worker_export_heatmaps,
-            (
-                i,
-                len(dfs_transponder),
-                dfs_transponder[i],
-                n_top_traces,
-                N_TRANSPOND,
-                ss_reshaped[:, : -N_TRANSPOND - 1],
-                dfs_ss[i].iloc[:, : -n_top_traces - 2],
-                timestamp[0],
-                THRESH_DT,
-                out,
-            ),
-        )
+    if miss_rate == 0:
+        pool = Pool(processes=args.n_job)
+        for i in range(len(dfs_transponder)):
+            result = pool.apply_async(
+                worker_export_heatmaps,
+                (
+                    i,
+                    len(dfs_transponder),
+                    dfs_transponder[i],
+                    n_top_traces,
+                    N_TRANSPOND,
+                    ss_reshaped[:, : -N_TRANSPOND - 1],
+                    dfs_ss[i].iloc[:, : -n_top_traces - 2],
+                    timestamp[0],
+                    THRESH_DT,
+                    out,
+                ),
+            )
 
-        xaxix_label, yaxis_label = result.get()[0], result.get()[1]
-    pool.close()
-    pool.join()
-    pool.terminate()
+            xaxix_label, yaxis_label = result.get()[0], result.get()[1]
+        pool.close()
+        pool.join()
+        pool.terminate()
 
     # for i in range(len(dfs_transponder)):
     #     d_t = dfs_transponder[i].iloc[:, :-n_top_traces - 2]
@@ -856,19 +857,20 @@ def main(args, raw_data, original_data_x, ids, timestamp, date_str, ss_data):
     # fig.write_html(filename)
 
     if export_csv:
-        export_imputed_data(
-            out,
-            data_m_x,
-            data_x,
-            imputed_data_x,
-            imputed_data_x_li,
-            timestamp,
-            date_str,
-            ids,
-            alpha,
-            hint_rate,
-            args.n_job,
-        )
+        if miss_rate == 0:
+            export_imputed_data(
+                out,
+                data_m_x,
+                data_x,
+                imputed_data_x,
+                imputed_data_x_li,
+                timestamp,
+                date_str,
+                ids,
+                alpha,
+                hint_rate,
+                args.n_job,
+            )
 
     # if export_traces:
     #   plot_imputed_data(out, imputed_data_x, imputed_data_x_li, raw_data, original_data_x, ids, timestamp)
